Update invalid structure.
