--- tests/invalid_toplevel_keys/duplicate_key/files/.platform.app.yaml
@@ -1,11 +1,0 @@
-name: app
-type: nodejs:22
-hooks:
-    build: |
-        npm install
-web:
-    commands:
-        start: npm start
-
-relationships:
-    db:
