--- tests/invalid_toplevel_keys/invalid_key/files/.platform/services.yaml
@@ -1,3 +1,0 @@
-db:
-    type: mariadb:10.4
-    disk: 1056
