+++ upsunvalidator/validate/upsun.py
@@ -2,9 +2,6 @@ import yaml
 import ruamel.yaml
 from ruamel.yaml.constructor import DuplicateKeyError
 
-# import ruamel.yaml
-# from ruamel.yaml.constructor import DuplicateKeyError
-
 # import sys
 # sys.tracebacklimit=0
